Upgraded to v1.2 & removed pm2 dependency
I removed the pm2 dependency since it's not required to run pm2 instances and is a severe security vulnerability. Along with dependency updates since the previous commit.

I also expanded the sslcheck functionality to run as a cronjob daily and post an embed if the days remaining is less than 7.