v1.3.2
Tweaked the date format code since the SSL cert had an extra space in the valid_to field which broke the old version of the code.

I also added another condition which checks if the days remaining on the SSL cert are between 8 & 14 with a more mild message vs. less than 7.